Split parser into multiple modules for clarity and improve documentation
