+++ src/parser/mod.rs
@@ -1,17 +1,6 @@
-use std::{
-    fmt,
-    num::{ParseFloatError, ParseIntError},
-};
+use std::num::{ParseFloatError, ParseIntError};
 
-use nom::{
-    self,
-    branch::alt,
-    bytes::complete::{tag, take_while, take_while1},
-    combinator::opt,
-    error::{context, VerboseError},
-    sequence::separated_pair,
-    IResult,
-};
+use nom::{self, error::VerboseError, IResult};
 use pest::{
     error::LineColLocation,
     iterators::{Pair, Pairs},
@@ -19,11 +8,13 @@ use pest::{
 };
 use thiserror::Error;
 
-use crate::ast::{statement::Statement, value::Value};
+use crate::ast::statement::Statement;
 
-pub use self::input::Input;
+pub use self::{input::Input, spanned::Spanned};
 
 pub mod input;
+pub mod literal;
+pub mod spanned;
 
 #[derive(Parser)]
 #[grammar = "parser/alloy.pest"]
@@ -41,22 +32,6 @@ pub enum ParserErrorKind {
 
 type ParserResult<'a, T> = IResult<Input<'a>, T, VerboseError<Input<'a>>>;
 type SpannedResult<'a, T> = ParserResult<'a, Spanned<T>>;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Spanned<T> {
-    pub ast: T,
-    pub start: usize,
-    pub end: usize,
-}
-
-impl<T> PartialEq<T> for Spanned<T>
-where
-    T: PartialEq,
-{
-    fn eq(&self, other: &T) -> bool {
-        self.ast == *other
-    }
-}
 
 #[derive(Debug)]
 pub struct ParserError {
@@ -131,458 +106,5 @@ pub fn parse(input: &str) -> Result<Vec<Statement>, ParserError> {
             kind: ParserErrorKind::WIP,
             location: e.line_col,
         }),
-    }
-}
-
-/// Parses a string into an `ast::Value::False` or `ast::Value::True`
-/// or returns an error.
-///
-/// # Examples
-/// ```
-/// use alloy::{ast::value::Value, parser::parse_bool};
-///
-/// let (input, value) = parse_bool("true".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::True);
-///
-/// let (input, value) = parse_bool("false".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::False);
-/// ```
-pub fn parse_bool(input: Input<'_>) -> SpannedResult<'_, Value> {
-    let start = input.position;
-    let (next_input, res) = context("bool", alt((tag("true"), tag("false"))))(input)?;
-    let value = if res == "true" {
-        Value::True
-    } else {
-        Value::False
-    };
-    let spanned = Spanned {
-        ast: value,
-        start,
-        end: next_input.position,
-    };
-    Ok((next_input, spanned))
-}
-
-/// Parse consecutive whitespaces including newline and carriage return
-/// characters.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::parser::parse_whitespace;
-///
-/// let (input, whitespace) = parse_whitespace(" \t\r\n123".into()).unwrap();
-/// assert_eq!(input, "123");
-/// assert_eq!(whitespace, " \t\r\n");
-/// ```
-pub fn parse_whitespace(input: Input<'_>) -> ParserResult<'_, Input<'_>> {
-    let (input, whitespace) = context(
-        "whitespace",
-        take_while(|p: char| p.is_whitespace() || p == '\n' || p == '\r'),
-    )(input)?;
-    Ok((input, whitespace))
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Sign {
-    Positive,
-    Negative,
-}
-
-impl Default for Sign {
-    fn default() -> Self {
-        Self::Positive
-    }
-}
-
-impl fmt::Display for Sign {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Positive => write!(f, "+"),
-            Self::Negative => write!(f, "-"),
-        }
-    }
-}
-
-/// Parse sign of a number either `+` or `-` into `Sign`.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::parser::{parse_sign, Sign};
-///
-/// let (input, sign) = parse_sign("+".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(sign, Sign::Positive);
-///
-/// let (input, sign) = parse_sign("-".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(sign, Sign::Negative);
-/// ```
-///
-pub fn parse_sign(input: Input<'_>) -> SpannedResult<Sign> {
-    let start = input.position;
-    let (next_input, sign) = context("sign", alt((tag("+"), tag("-"))))(input)?;
-    let sign = if sign == "+" {
-        Sign::Positive
-    } else {
-        Sign::Negative
-    };
-    let spanned = Spanned {
-        ast: sign,
-        start,
-        end: next_input.position,
-    };
-    Ok((next_input, spanned))
-}
-
-/// Parse one or more digits with given radix and underscores can be used
-/// for improved readability for large constants.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::parser::parse_digits;
-///
-/// // Parse decimal digits (base/radix 10)
-/// let (input, digits) = parse_digits("123".into(), 10).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(digits, 123);
-///
-/// // Parse decimal digits (base/radix 10)
-/// let (input, digits) = parse_digits("1_000_000".into(), 10).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(digits, 1_000_000);
-///
-/// // Parse hexadecimal digits (base/radix 16)
-/// let (input, digits) = parse_digits("FF12a".into(), 16).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(digits, 1044778);
-/// ```
-///
-/// # Errors
-///
-/// This function will return an error if given input doesn't contain digits of given radix.
-pub fn parse_digits(input: Input<'_>, radix: u32) -> ParserResult<'_, i64> {
-    let (input, digits) = context(
-        "digits",
-        take_while1(|c: char| c.is_digit(radix) || c == '_'),
-    )(input)?;
-    if digits.input.starts_with("_") {
-        todo!("parse_digits: handle underscores");
-    }
-    let number = i64::from_str_radix(&digits.input.replace("_", ""), radix).unwrap();
-    Ok((input, number))
-}
-
-/// Parse decimal integer into `i64` and convert it to `Value::Integer`.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::{ast::value::Value, parser::parse_decimal};
-///
-/// let (input, value) = parse_decimal("123".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(123));
-///
-/// let (input, value) = parse_decimal("- 123".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(-123));
-///
-/// ```
-///
-/// # Errors
-///
-/// This function will return an error if doesn't contain decimal digits.
-pub fn parse_decimal(input: Input<'_>) -> SpannedResult<'_, Value> {
-    parse_radix_integer(input, 10, None)
-}
-
-/// Parse hexadecimal integer into `i64` and convert it to `Value::Integer`.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::{ast::value::Value, parser::parse_hexadecimal};
-///
-/// let (input, value) = parse_hexadecimal("0xFF".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(255));
-///
-/// let (input, value) = parse_hexadecimal("0xab".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(171));
-///
-/// ```
-///
-/// # Errors
-///
-/// This function will return an error if it contains doesn't start with
-/// binary prefix `0x` or contains invalid hexadecimal digits.
-pub fn parse_hexadecimal(input: Input<'_>) -> SpannedResult<'_, Value> {
-    parse_radix_integer(input, 16, Some("0x"))
-}
-
-/// Parse octal integer into `i64` and convert it to `Value::Integer`.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::{ast::value::Value, parser::parse_octal};
-///
-/// let (input, value) = parse_octal("0o77".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(63));
-///
-/// let (input, value) = parse_octal("0o11".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(9));
-///
-/// ```
-///
-/// # Errors
-///
-/// This function will return an error if it contains doesn't start with
-/// binary prefix `0o` or contains invalid octal digits.
-pub fn parse_octal(input: Input<'_>) -> SpannedResult<'_, Value> {
-    parse_radix_integer(input, 8, Some("0o"))
-}
-
-/// Parse binary number into `i64` and convert it to `Value::Integer`.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::{ast::value::Value, parser::parse_binary};
-///
-/// let (input, value) = parse_binary("0b101".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(5));
-///
-/// let (input, value) = parse_binary("0b1001".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(9));
-///
-/// ```
-///
-/// # Errors
-///
-/// This function will return an error if it contains doesn't start with
-/// binary prefix `0b` or contains invalid binary digits.
-pub fn parse_binary(input: Input<'_>) -> SpannedResult<'_, Value> {
-    parse_radix_integer(input, 2, Some("0b"))
-}
-
-/// Parse hexadecimal integer into `i64` and convert it to `Value::Integer`.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::{ast::value::Value, parser::parse_radix_integer};
-///
-/// let (input, value) = parse_radix_integer("0xFF".into(), 16, "0x".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(255));
-///
-/// let (input, value) = parse_radix_integer("0b101".into(), 2, "0b".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(value, Value::Integer(5));
-///
-/// ```
-///
-/// # Errors
-///
-/// This function will return an error if it contains doesn't start with given prefix
-/// or contains invalid digits for given radix.
-pub fn parse_radix_integer<'a>(
-    input: Input<'a>,
-    radix: u32,
-    prefix: Option<&'a str>,
-) -> SpannedResult<'a, Value> {
-    let start = input.position;
-
-    // Parse sign of the number or default to positive
-    let (input, sign) = context("radix integer", opt(parse_sign))(input)?;
-    let sign = if let Some(sign) = sign {
-        sign.ast
-    } else {
-        Sign::default()
-    };
-
-    // Any number of whitespace characters can follow the sign
-    let (input, _) = parse_whitespace(input)?;
-
-    // and raidx is not 10, then we need to parse the prefix
-    let input = if let Some(prefix) = prefix {
-        let (input, _) = context("radix integer prefix", tag(prefix))(input)?;
-        input
-    } else {
-        input
-    };
-
-    // FIXME: Instead of unwrapping result here, we should return an error
-    let (input, integer) = context("radix integer", |input| parse_digits(input, radix))(input)?;
-    let integer = match sign {
-        Sign::Positive => Value::Integer(integer),
-        Sign::Negative => Value::Integer(-integer),
-    };
-
-    let spanned = Spanned {
-        ast: integer,
-        start,
-        end: input.position,
-    };
-    Ok((input, spanned))
-}
-
-/// Parse integer number not limited to decimal including binary, octal, hexadecimal.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::parser::parse_integer;
-///
-/// assert!(parse_integer("124".into()).is_ok());
-/// assert!(parse_integer("0b111".into()).is_ok());
-/// assert!(parse_integer("0o111".into()).is_ok());
-/// assert!(parse_integer("0xb1AF".into()).is_ok());
-/// ```
-///
-pub fn parse_integer(input: Input<'_>) -> SpannedResult<'_, Value> {
-    context(
-        "integer",
-        alt((parse_decimal, parse_hexadecimal, parse_octal, parse_binary)),
-    )(input)
-}
-
-fn parse_decimal_digits(input: Input<'_>) -> ParserResult<'_, i64> {
-    parse_digits(input, 10)
-}
-
-fn parse_float_optional(input: Input<'_>) -> ParserResult<'_, (i64, i64)> {
-    let (input, (whole, fractional)) =
-        separated_pair(opt(parse_decimal_digits), tag("."), parse_decimal_digits)(input)?;
-    let whole = whole.unwrap_or_default();
-    Ok((input, (whole, fractional)))
-}
-
-fn parse_float_dot_optional(input: Input<'_>) -> ParserResult<'_, (i64, i64)> {
-    let (input, (whole, fractional)) =
-        separated_pair(parse_decimal_digits, tag("."), opt(parse_decimal_digits))(input)?;
-    let fractional = fractional.unwrap_or_default();
-    Ok((input, (whole, fractional)))
-}
-
-fn fractional_part(mut float: f64) -> f64 {
-    while float > 1.0 {
-        float /= 10.0;
-    }
-    float
-}
-
-fn float_from_parts(sign: Sign, whole: i64, fractional: i64) -> Value {
-    let float = whole as f64 + fractional_part(fractional as f64);
-    let float = match sign {
-        Sign::Positive => float,
-        Sign::Negative => -float,
-    };
-    Value::Float(float)
-}
-
-/// Parse floating point number.
-///
-/// # Examples
-///
-/// ```
-/// use alloy::{ast::value::Value, parser::parse_float};
-///
-/// let (input, float) = parse_float("1.23".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(float, Value::Float(1.23));
-///
-/// let (input, float) = parse_float("145.15".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(float, Value::Float(145.15));
-///
-/// let (input, float) = parse_float("- 145.15".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(float, Value::Float(-145.15));
-///
-/// let (input, float) = parse_float(".15".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(float, Value::Float(0.15));
-///
-/// let (input, float) = parse_float("5.".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(float, Value::Float(5.0));
-///
-/// let (input, float) = parse_float("5_000.600_600".into()).unwrap();
-/// assert_eq!(input, "");
-/// assert_eq!(float, Value::Float(5000.6006));
-/// ```
-///
-/// # Errors
-///
-/// This function will return an error if input doesn't contain a valid floating point number.
-pub fn parse_float(input: Input<'_>) -> SpannedResult<'_, Value> {
-    let start = input.position;
-    let (input, sign) = context("float", opt(parse_sign))(input)?;
-    let sign = if let Some(sign) = sign {
-        sign.ast
-    } else {
-        Sign::default()
-    };
-    let (input, _) = parse_whitespace(input)?;
-    let (input, (whole, fractional)) = context(
-        "float",
-        alt((parse_float_optional, parse_float_dot_optional)),
-    )(input)?;
-    let float = float_from_parts(sign, whole, fractional);
-    let spanned = Spanned {
-        ast: float,
-        start,
-        end: input.position,
-    };
-    Ok((input, spanned))
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        ast::value::Value,
-        parser::{parse_sign, Sign},
-    };
-
-    use super::parse_bool;
-
-    #[test]
-    fn test_boolean() {
-        let (rest, spanned) = parse_bool("true".into()).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(spanned.ast, Value::True);
-        assert_eq!(spanned.start, 0);
-        assert_eq!(spanned.end, 4);
-        let (rest, spanned) = parse_bool("false".into()).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(spanned.ast, Value::False);
-        assert_eq!(spanned.start, 0);
-        assert_eq!(spanned.end, 5);
-    }
-
-    #[test]
-    fn test_sign() {
-        let (rest, spanned) = parse_sign("+".into()).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(spanned.ast, Sign::Positive);
-        assert_eq!(spanned.start, 0);
-        assert_eq!(spanned.end, 1);
-        let (rest, spanned) = parse_sign("-".into()).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(spanned.ast, Sign::Negative);
-        assert_eq!(spanned.start, 0);
-        assert_eq!(spanned.end, 1);
     }
 }
